+++ BatScript/errorsys.h
@@ -7,7 +7,7 @@ namespace Bat
 	class ErrorSys
 	{
 	public:
-		static void Report( int line, int column, const std::string& message );
+		static void Report( size_t line, size_t column, const std::string& message );
 		static bool HadError();
 		static void SetSource( const std::string& source );
 	};
